fix su gestione utenti

+++ corso-angular/express-test/routes/users.js
@@ -2,6 +2,7 @@ var express = require('express');
 var router = express.Router();
 var db = require('../db');
 var userDao = require('../dao/user.dao.js');
+var cors = require('cors');
 
 router.get('/', function(req, res) {
 
@@ -9,6 +10,7 @@ router.get('/', function(req, res) {
     res.json(docs);
   });
 });
+router.options('/',cors({credentials: true, origin: '*',preflightContinue:true}));
 
 router.get('/:id', function(req, res) {
   userDao.getUserByUid(req.params.id, function(err, docs) {
@@ -24,13 +26,14 @@ router.get('/:id', function(req, res) {
 });
 
 router.post('/', function(req, res) {
-  user.getUserByUid(req.body.id, function(err, doc) {
+  var user=req.body;
+  userDao.getUserByUid(user.uid, function(err, doc) {
     if (doc) {
       res.status(409).json({
         error: 'Event already present.',
       });
     } else {
-      userDao.save(res.body, function(err, doc) {
+      userDao.save(user, function(err, doc) {
         if (err) {
           res.status(500).json(err);
         } else {
@@ -63,10 +66,16 @@ router.put('/:id', function(req, res) {
 });
 
 router.delete('/:id', function(req, res) {
-
+  console.log(req.params.id);
+  if(!req.user.isAdmin){
+    return res.status(403).send({
+      success: false,
+      message: 'You aren\'t Administrator.',
+    });
+  }
   var collection = db.get().collection('users');
   collection.findOneAndDelete({
-      id: parseInt(req.params.id),
+      uid: req.params.id,
     }, {
       projection: {
         _id: 0,
